Struct uxi::Client

pub struct Client<C: Send> { /* private fields */ }

Client represents an UXI engine client. It can accept and parse commands from the GUI and send commands to the GUI though its input and output. Commands sent from the GUI are automatically parsed and executed according to the Command schema provided by the user to the Client. The client supports any UXI type protocol, including but not limited to UCI and UAI.

Options can also be added to a Client in the form of parameters. See the documentation of Parameter and the Client::option function for more details.

Implementations

impl<C: Send + Default + 'static> Client<C>

pub fn start(&self)

start starts the Client so that it can now accept Commands from the GUI and send Commands back to the GUI as necessary. The Client will return only if it encounters a fatal error while running a command (RunError::Fatal) or one of the commands asks the Client to quit (RunError::Quit).

pub fn run_cmd_string(&self, str: String) -> CmdResult

run_cmd_strings allows running a Command independently from the main uxi loop. Since the commands are run in a standalone way, everything is run synchronously.

impl<C: Send> Client<C>

pub fn new() -> Self

new creates a new Client. The Client can be configured using builder methods like Client::command, Client::option, etc. These functions take the ownership of the given Client value and return that ownership after adding the given option. This allows the usage of the builder pattern while creating a new Client.

let client = Client::new()
    .command("go", go_cmd)
    .command("perft", perft_cmd);

pub fn command(self, name: &str, cmd: Command<C>) -> Self

command adds the given Command to the given Client. After this, the Client will be able to parse and run that Command when such a request is sent from the GUI.

let client = Client::new()
    .command("go", go_cmd)
    .command("perft", perft_cmd);

pub fn option(self, name: &str, option: Parameter) -> Self

option adds the given option to the given Client. This will register that option with the Client, and those options can then be set by GUIs using the engine with a UXI type protocol. The options will also be made available in the context so that Commands can retrieve and use their values.

let client = Client::new()
    .option("Hash", Parameter::Spin(16, 1, 33554432))
    .option("Threads", Parameter::Spin(1, 1, 1024));

pub fn protocol(self, name: &str) -> Self

protocol configures the Client to support the given UXI protocol. UXI protocols have a name in the format “uxi”, where the x can be replaced by any lowercase alphabet, usually the leading letter of the name of the game played by the engine.

Some popular protocols are “uci” for Chess and “uai” for Ataxx. The name “ugi” however is reserved and may not be used as a protocol name. This is because ugi is the identifier for the Universal Game Interface, which is a game-agnostic protocol which is supported by all Clients by default.

let client = Client::new()
    .protocol("uci");

pub fn engine(self, name: &str) -> Self

engine sets the name of the engine for the Client.

This information is reported to the GUI when the uxi Command is invoked, and is also accessible by Commands through their bundles.

let client = Client::new()
    .engine("Stockfish2");

pub fn author(self, name: &str) -> Self

author sets the name of the author of the Client’s engine.

This information is reported to the GUI when the uxi Command is invoked, and is also accessible by Commands through their bundles.

let client = Client::new()
    .author("Rak Laptudirm");