Struct uxi::Command

pub struct Command<C: Send> { /* private fields */ }

Command represents a runnable UXI command. It contains all the metadata needed to parse and verify a Command request from the GUI for a Command, and to run that Command with the current context and the provided flag values. The C type parameter is the type of the context used by Client to maintain its state across different command invocations and to synchronize threads.

A Command’s schema is composed of its name, run function, flag schema, and whether it is run in parallel. When a Command is invoked by a GUI, the invocation starts with its name followed by any number of flags from its flag schema. See the documentation of Flag for more details.

Implementations

impl<C: Send> Command<C>

pub fn new(func: RunFn<C>) -> Command<C>

new creates a new Command with the given run function.

By default the flag schema is empty the the Command is run synchronously.

Further configuration of the Command can be done using the builder style methods provided on Command, like Self::flag and Self::parallelize. These methods take the ownership of the given Command, make the necessary changes and then return it. These allows them to be chained in builder pattern style to create fully configured Commands.

let cmd: Command<C> =
    // new invocation to create a Command. In this example, a very
    // simple run function which returns `Ok(())` is provided.
    Command::new(|_ctx, _flg| Ok(()))
        // Add flags to the Command's flag schema.
        .flag("flag1", Flag::Boolean)
        .flag("flag2", Flag::Singular)
        .flag("flag3", Flag::Array(10))
        .flag("flag4", Flag::Variadic)
        // Make the command run in parallel.
        .parallelize(true);

pub fn flag(self, name: &str, flag: Flag) -> Self

flag adds the given Flag to the Command’s Flag schema.

After a flag is added to a Command, the Client will automatically parse that flag and verify its arguments whenever that Command is invoked.

flag is a builder method. Refer to the documentation of Command::new for more information about Command’s builder methods.

pub fn parallelize(self, y: bool) -> Self

parallelize sets if the Command to be run in a separate thread.

Synchronization of a parallel Command should be done with the help of the mutex-locked context that is provided to the Command’s run function. The Client continues in the main thread while a parallel Command is running.

parallelize is a builder method. Refer to the documentation of Command::new for more information about Command’s builder methods.